refactored the code added, polyglot scripts, fixed server streaming example
