--- src/main/proto/Blog/Blog.proto
@@ -1,62 +1,0 @@
-syntax="proto3";
-
-package blogDemo;
-
-option java_package= "com.proto.blog";
-option java_multiple_files = true;
-
-message Blog {
-    string id = 1;
-    string author_id = 2;
-    string title = 3;
-    string content = 4;
-}
-
-message CreateBlogRequest {
-    Blog blog = 1; // the id shouldn't matter
-}
-
-message CreateBlogResponse {
-    Blog blog = 1; // the id will correspond to the one in MongoDB
-}
-
-message ReadBlogRequest {
-    string blog_id = 1;
-}
-
-message ReadBlogResponse {
-    Blog blog = 1;
-}
-
-message UpdateBlogRequest {
-    Blog blog = 1;
-}
-
-message UpdateBlogResponse {
-    Blog blog = 1;
-}
-
-message DeleteBlogRequest {
-    string blog_id = 1;
-}
-
-message DeleteBlogResponse {
-    string blog_id = 1;
-}
-
-message ListBlogRequest {
-
-}
-
-message ListBlogResponse {
-    Blog blog = 1;
-}
-
-service BlogService {
-    rpc CreateBlog(CreateBlogRequest) returns (CreateBlogResponse){};
-    rpc ReadBlog(ReadBlogRequest) returns (ReadBlogResponse){}; // return NOT_FOUND if not found
-    rpc UpdateBlog(UpdateBlogRequest) returns (UpdateBlogResponse){}; // return NOT_FOUND if not found
-    rpc DeleteBlog(DeleteBlogRequest) returns (DeleteBlogResponse){}; // return NOT_FOUND if not found
-    rpc ListBlog(ListBlogRequest) returns (stream ListBlogResponse){};
-}
-
