--- src/main/proto/Sample/Greet.proto
@@ -1,37 +1,0 @@
-syntax = "proto3";
-
-package greet;
-option java_package = "com.proto.greet";
-option java_multiple_files = true;
-
-message Greeting {
-    string first_name = 1;
-    string last_name = 2;
-}
-message GreetRequest {
-    Greeting greeting = 1;
-}
-message GreetResponse {
-    string result = 1;
-}
-message GreetManyTimesRequest {
-    Greeting greeting = 1;
-}
-message GreetManyTimesResponse {
-    string result = 1;
-}
-message LongGreetRequest {
-    Greeting greeting = 1;
-}
-message LongGreetResponse {
-    string result = 1;
-}
-
-service GreetService {
-    //Unary
-    rpc Greeter(GreetRequest) returns (GreetResponse) {};
-    //Server streaming
-    rpc GreetManyTimes(GreetManyTimesRequest) returns (stream GreetManyTimesResponse) {};
-    //client streaming
-    rpc LongGreet(stream LongGreetRequest) returns (LongGreetResponse) {};
-}
