--- src/main/proto/Sample/Sample.proto
@@ -1,9 +1,0 @@
-syntax = "proto3";
-
-package sampleGrpc;
-
-option java_package = "com.proto.sample";
-
-message SampleMessage {}
-
-service SampleService {}
