--- src/main/proto/bidirectionalStreaming/Chat.proto
@@ -1,34 +1,0 @@
-/*
- * Copyright 2016 Google, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-syntax = "proto3";
-
-import "google/protobuf/timestamp.proto";
-package com.example.grpc.chat;
-
-message ChatMessage {
-    string from = 1;
-    string message = 2;
-}
-
-message ChatMessageFromServer {
-    google.protobuf.Timestamp timestamp = 1;
-    ChatMessage message = 2;
-}
-
-service ChatService {
-    rpc chat(stream ChatMessage) returns (stream ChatMessageFromServer);
-}
